requirements.txt
PyGithub==2.1.1
sqlite3-api==2.0.1
python-dotenv==1.0.0
